Fix printing of Vararg on Julia 1.7
Recent versions of Julia has moved Vararg from being a Type to being its own
thing. This means it will error if T is a vararg and you try to do `isa` to `<:`
operations with it.
This caused an error, which is fixed here by explicitly testing if T is a Vararg.